+++ tests/test_version.py
@@ -1,8 +1,6 @@
-import pytest
-
 # verify that the version is correct
 from cloudwatcher._version import __version__
 
 
 def test_version():
-    assert __version__ == "0.0.4"
+    assert __version__ == "0.0.5"